Mobile/feature/home

+++ server/src/main/java/com/csse/server/model/Item.java
@@ -9,15 +9,15 @@ public class Item {
     @Id
     private ObjectId id;
     private String name;
-    private Map<String,Float> supplier;
+    private Map<String,Double> supplier;
     private String chosenOne;
     private String chosenOnesPrice;
 
     private String description;
     private boolean restricted;
-    private float restrictedAmount;
+    private double restrictedAmount;
 
-    public Item(String name, Map<String,Float> supplier, String description, boolean restricted, float restrictedAmount) {
+    public Item(String name, Map<String,Double> supplier, String description, boolean restricted, double restrictedAmount) {
         this.name = name;
         this.supplier = supplier;
         this.description = description;
@@ -39,11 +39,11 @@ public class Item {
         this.name = name;
     }
     
-    public Map<String,Float> getSupplier() {
+    public Map<String,Double> getSupplier() {
         return supplier;
     }
 
-    public void setSupplier(Map<String,Float> supplier) {
+    public void setSupplier(Map<String,Double> supplier) {
         this.supplier = supplier;
     }
 
@@ -63,11 +63,11 @@ public class Item {
         this.restricted = restricted;
     }
 
-    public float getRestrictedAmount() {
+    public double getRestrictedAmount() {
         return restrictedAmount;
     }
 
-    public void setRestrictedAmount(float restrictedAmount) {
+    public void setRestrictedAmount(double restrictedAmount) {
         this.restrictedAmount = restrictedAmount;
     }
